requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0